fix: fix uint16

+++ contracts/0.8.9/oracle/ValidatorsExitBusOracle.sol
@@ -50,8 +50,8 @@ contract ValidatorsExitBusOracle is BaseOracle, ValidatorsExitBus {
         uint256 genesisTime,
         address lidoLocator,
         address nodeOperatorsRegistry,
-        uint32 maxBalanceWcType01Eth,
-        uint32 maxBalanceWcType02Eth
+        uint16 maxBalanceWcType01Eth,
+        uint16 maxBalanceWcType02Eth
     ) BaseOracle(secondsPerSlot, genesisTime) ValidatorsExitBus(lidoLocator, nodeOperatorsRegistry, maxBalanceWcType01Eth, maxBalanceWcType02Eth) {}
 
     function initialize(
